requirements.txt
pymodbus==3.6.8
pyyaml
